requirements.txt
streamlit
boto3
pandas
openpyxl
python-dotenv
pydantic
openai
